FIX ex01 to ex09 mdr

+++ day00/ex04/whatis.py
@@ -2,7 +2,7 @@ import sys
 
 def checknumber(object: str) -> bool:
     try:
-        float(object)
+        int(object)
         return (True)
     except:
         return (False)
